Extract QuadkeyTileset2D into separate class
